First commit of the Repo. First designs of modules.
This is the first commit of the repo, where the modules of the
RAM, ROM and a multiplexer to select the output of the two memories.
These modules are working, and a .hex file has been created to store
the information that will be loaded in the ROM.

Andres Rivera Marquez.